feat: Enable dashboard functionality
- Added logic for the dashboard.
- Implemented data handling to manage dashboard data efficiently.
- This includes enhancements to the user interface and data handling for a better user experience.

+++ lib/core/utils/widgets/adaptive_layout.dart
@@ -11,7 +11,7 @@ class AdaptiveLayout extends StatelessWidget {
       this.mobileMaxWidth = 760,
       this.tabletMaxWidth = 1200});
   final WidgetBuilder mobileLayout, tabletLayout, desktopLayout;
-  final int mobileMaxWidth, tabletMaxWidth;
+  final double mobileMaxWidth, tabletMaxWidth;
 
   @override
   Widget build(BuildContext context) {
